requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
requests==2.31.0
langchain==0.1.0
langchain-community==0.0.10
faiss-cpu==1.7.4
sentence-transformers==2.2.2
pdf2image==1.16.3
python-docx==0.8.11
docx2txt==0.8
pypdf==3.17.1
torch==2.1.0
transformers==4.36.2
aiohttp==3.9.1